[refactor] move the signaling service to /signal
